WIP: 59 failing tests
* No more shiro, it doesn't work well with spring-boot 2.0
* Ideally we'd like to have a single profile (application-unit-test-permissions)
with one user that has full permissions on /* and all of our tests can
use it. It would force us to be consistent across every test (as opposed
to giving random users permissions). Unfortunately, the database is
seeded with permissions from the profile before each test file, but they
are clobbered after most tests. We can use @DirtiesContext to seed the
database before each test but it makes the tests significantly slower

[#157596194] Operators can add permissions to credentials that do not yet exist using properties in the manifest